// src/lib/types.ts
export interface Emprendimiento {
  id: string;
  name: string;
  description: string;
  image: string;
  imageHint: string;
  goal: number;
  raised: number;
  whatsappMessage: string;
}

export interface VoluntarioRole {
  title: string;
  description: string;
  image: string;
  imageHint: string;
}